requirements.txt
pdfkit
streamlit
snowflake-connector-python
